% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ritable-package.R
\docType{package}
\name{Ritable-package}
\alias{Ritable}
\alias{Ritable-package}
\title{Ritable: Pedigree-Based Heritability Estimation for Family Cohort Studies}
\description{
Provides profile-likelihood variance-components estimation of narrow-sense heritability (h2) for quantitative traits in family cohort studies. Additive genetic relationship matrices are built from pedigrees via 'kinship2'. Phenotypes are inverse-normal transformed internally. Likelihood-ratio tests use a one-sided chi-squared boundary correction equivalent to SOLAR Eclipse. Ninety-five percent confidence intervals are derived from the profile likelihood rather than Wald approximations. Batch estimation over many traits returns tidy data frames ready for downstream visualisation (forest plots, heatmaps).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-itable.circadia-lab.uk}
  \item \url{https://github.com/circadia-bio/R-itable}
  \item Report bugs at \url{https://github.com/circadia-bio/R-itable/issues}
}

}
\author{
\strong{Maintainer}: Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})

Authors:
\itemize{
  \item Mario Leocadio-Miguel \email{mario.miguel@northumbria.ac.uk} (\href{https://orcid.org/0000-0002-7248-3529}{ORCID})
}

}
\keyword{internal}
